fix: exports GizmoLog<N>
to fix cannot be named issue
