+++ frameworks/log/ts/index.ts
@@ -1,5 +1,5 @@
 export * from './log_gizmo.js'
-export type { LogGizmoOptions } from './log_gizmo.types.js'
+export type { LogGizmoOptions, GizmoLog } from './log_gizmo.types.js'
 export * from './logger.js'
 export * from './old_log.js'
 export * from './standard_log.js'
